requirements.txt
Cython==0.29.33
python-dotenv==0.21.0
PyYAML==6.0
